+++ packages/iov-rise/src/riseconnection.ts
@@ -65,7 +65,7 @@ async function loadChainId(baseUrl: string): Promise<ChainId> {
   const url = checkAndNormalizeUrl(baseUrl) + "/api/blocks/getNethash";
   const result = await axios.get(url);
   const responseBody = result.data;
-  return responseBody.nethash;
+  return `rise-${responseBody.nethash.slice(0, 10)}` as ChainId;
 }
 
 export class RiseConnection implements BlockchainConnection {
@@ -80,8 +80,10 @@ export class RiseConnection implements BlockchainConnection {
   public constructor(baseUrl: string, chainId: ChainId) {
     this.baseUrl = checkAndNormalizeUrl(baseUrl);
 
-    if (!chainId.match(/^[a-f0-9]{64}$/)) {
-      throw new Error("The chain ID must be a RISE nethash, encoded as 64 lower-case hex characters.");
+    if (!chainId.match(/^rise-[a-f0-9]{10}$/)) {
+      throw new Error(
+        "The chain ID must be a Rise nethash, encoded as `rise-%s` where `%s` is the 10-digit hex-encoded prefix of the relevant nethash.",
+      );
     }
     this.myChainId = chainId;
   }